backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
python-dotenv==1.2.1
markitdown==0.1.4
pypdfium2==5.3.0
python-docx==1.2.0
pillow==12.1.0
groq==1.0.0
pinecone>=3.1.0
google-auth==2.47.0
google-auth-oauthlib==1.2.4
slowapi==0.1.9
redis==5.0.1
celery==5.3.6